job monitor-karkidi/requirements.txt
# Web Scraping
selenium
beautifulsoup4
requests
lxml
html5lib

# Data Processing
pandas
numpy
openpyxl

# Machine Learning
scikit-learn
nltk
spacy
transformers
torch

# Text Processing
textblob
fuzzywuzzy
python-levenshtein

# Web Framework & Dashboard
streamlit
plotly
altair

# Database
sqlite3

# Email & Notifications
smtplib
email-validator

# Configuration & Environment
python-dotenv
pyyaml
configparser

# Logging & Monitoring
loguru

# Utilities
python-dateutil
pytz
schedule
tqdm

# Development & Testing
pytest
pytest-cov
black
flake8
pre-commit

# Deployment
gunicorn
uvicorn